+++ lib/rspec/mocks/standalone.rb
@@ -1,2 +1,3 @@
 require 'rspec/mocks'
 include RSpec::Mocks::ExampleMethods
+RSpec::Mocks.setup